refactor: Organize files
