+++ src/components/Card.js
@@ -1,19 +1,20 @@
 import React from 'react';
 import './Card.css';
 
-const Card = ({person, colors}) => {
-  const {first, last} = person.name;
+const Card = ({ person, colors }) => {
+  const { first, last } = person.name;
   const imgSrc = person.picture.large;
 
-  
   const backgroundGradient = {
-    backgroundImage: `linear-gradient(135deg, ${colors[0]} 0%, ${colors[1]} 100%)`,
-  }
+    backgroundImage: `linear-gradient(135deg, ${colors[0]} 0%, ${
+      colors[1]
+    } 100%)`,
+  };
 
   return (
     <div className="card">
       <div className="card-header" style={backgroundGradient}>
-        <img src={imgSrc} className="card-img"/>
+        <img src={imgSrc} className="card-img" />
         <div className="avatar">
           <h3 className="avatar-name">{`${first} ${last}`}</h3>
           <p className="avatar-handle">{`@${first}${last}`}</p>
@@ -27,7 +28,7 @@ const Card = ({person, colors}) => {
         </div>
       </div>
     </div>
-  )
-}
+  );
+};
 
 export default Card;
